Implemented a new gallery plugin as well as custom jQuery
They jQuery implemented is a simple check of all the H3's on the page.
If they do not have a span in them already, then add one. This is
critical to their styling. This way, the client can create a new custom
page without the need for me for every new update.